requirements.txt
# DL
dm-haiku
chex==0.1.5
einops
optax

# Remote logging
wandb

# Data
datasets
tokenizers

click
pyyaml
tqdm

# CI and Dev-Ops
mypy
black
flake8
isort
pytest
